requirements.txt
ms-swift[llm]
transformers>=4.51.0
accelerate>=0.34.0
peft>=0.12.0
datasets>=2.21.0
pyyaml>=6.0.2